+++ CHANGES.rst
@@ -89,6 +89,9 @@ New Features
     attributes, methods or a numpy.ndarray-like interface like ``NDDataArray``.
     [#4797]
 
+  - ``NDIOMixin`` and subclasses now support reading/writing fits files using
+    the ``read`` or ``write`` method with ``format='simple_fits'``. [#4799]
+
 - ``astropy.stats``
 
   - Added ``axis`` keyword for ``mad_std`` function. [#4688]